fix : goal fsd 구조 적용
