+++ src/entities/goal/api/index.ts
@@ -4,3 +4,4 @@
 
 export * from './fetchGoal';
 export * from './updateGoalTitle';
+export * from './mocks';
