+++ src/entities/goal/api/mocks/index.ts
@@ -1,0 +1,2 @@
+export * from './goalHandlers';
+export * from './mocks';
