+++ src/entities/goal/index.ts
@@ -1,4 +1,2 @@
-// Goal 엔티티 Public API
 export * from './model';
-// export * from './api';
-// export * from './ui';
+export * from './api';
