+++ src/entities/goal/model/index.ts
@@ -1,2 +1,3 @@
 export * from './queryKeys';
 export * from './types';
+export * from './useGoal';
