--- src/entities/goal/ui/index.ts
@@ -1,4 +1,0 @@
-// Goal UI 세그먼트 Public API
-// export * from './GoalCard';
-// export * from './GoalList';
-// export * from './GoalForm';
