+++ src/entities/study/api/index.ts
@@ -1,7 +1,3 @@
-// Study API 세그먼트 Public API
-// export * from './study-api';
-// export * from './queries';
-
 export * from './fetchStudy';
 export * from './updateStudyInfo';
 export * from './mocks';
