+++ src/entities/study/index.ts
@@ -1,4 +1,3 @@
-// Study 엔티티 Public API
 export * from './model';
-// export * from './api';
-// export * from './ui';
+export * from './api';
+export * from './ui';
