+++ src/entities/study/ui/index.ts
@@ -1,4 +1,1 @@
-// Study UI 세그먼트 Public API
-// export * from './StudyCard';
-// export * from './StudyList';
-// export * from './StudyForm';
+export * from './studyInfo';
