+++ src/entities/todo/index.ts
@@ -1,4 +1,1 @@
-// Todo 엔티티 Public API
 export * from './model';
-// export * from './api';
-// export * from './ui';
