+++ src/entities/todo/model/index.ts
@@ -1,2 +1,1 @@
-export * from './queryKeys';
 export * from './types';
