--- src/entities/todo/model/queryKeys.ts
@@ -1,4 +1,0 @@
-export const todoQueryKeys = {
-  all: ['todo'] as const,
-  // list: (todoListId: string) => [...todoQueryKeys.all, 'list', todoListId] as const,
-} as const;
