--- src/features/Navigate_Note/index.ts
@@ -1,1 +1,0 @@
-export * from './ui';
