--- src/features/Navigate_Note/ui/index.ts
@@ -1,1 +1,0 @@
-export * from './navigate_note';
