--- src/features/Navigate_Todo/index.ts
@@ -1,1 +1,0 @@
-export * from './ui';
