--- src/features/Navigate_Todo/ui/index.ts
@@ -1,1 +1,0 @@
-export * from './navigate_todo';
