--- src/shared/lib/mocks/handlers/todoHandlers.ts
@@ -1,4 +1,0 @@
-// todo 관련 핸들러 필요시 여기에 추가
-export const todoHandlers = [
-  // 예시: todo 관련 GET, PATCH 등
-];
