+++ src/widgets/goal/index.ts
@@ -1,1 +1,1 @@
-// export * from './ui'
+export * from './ui';
