+++ src/widgets/goal/ui/index.ts
@@ -1,1 +1,1 @@
-// export * from './goal';
+export * from './goal';
